LRDOCS-8379 Adding Bar
